feat: Add recently played page
This commit adds a new `RecentlyPlayedPage` widget to display a list of recently played tracks. The widget receives a JSON string as input and parses it to extract the necessary data for each track. It then displays the track's album image, name, artist, album, and the date it was played. The page is wrapped in a `Scaffold` and includes an `AppBar` with the title "Recently Played". The list of tracks is displayed using a `ListView.builder` and each track is represented by a `ListTile` widget.

The purpose of this commit is to provide a way for users to view their recently played tracks in the app.

Note: This commit message follows the conventional format of "feat: <description>".